+++ myapp/settings/dev.py
@@ -1,10 +1,5 @@
 from .base import *
 
-import os
-import sys
-import json
-import base64
-
 # SECURITY WARNING: don't run with debug turned on in production!
 DEBUG = True
 
@@ -22,56 +17,3 @@ try:
 except ImportError:
     pass
 
-#################################################################################
-# Platform.sh-specific configuration
-
-# This variable should always match the primary database relationship name,
-#   configured in .platform.app.yaml.
-PLATFORMSH_DB_RELATIONSHIP="database"
-
-# Helper function for decoding base64-encoded JSON variables.
-def decode(variable):
-    """Decodes a Platform.sh environment variable.
-    Args:
-        variable (string):
-            Base64-encoded JSON (the content of an environment variable).
-    Returns:
-        An dict (if representing a JSON object), or a scalar type.
-    Raises:
-        JSON decoding error.
-    """
-    try:
-        if sys.version_info[1] > 5:
-            return json.loads(base64.b64decode(variable))
-        else:
-            return json.loads(base64.b64decode(variable).decode('utf-8'))
-    except json.decoder.JSONDecodeError:
-        print('Error decoding JSON, code %d', json.decoder.JSONDecodeError)
-
-# Import some Platform.sh settings from the environment.
-if (os.getenv('PLATFORM_APPLICATION_NAME') is not None):
-    DEBUG = False
-    # WAGTAILADMIN_BASE_URL
-    # BASE_URL
-    if (os.getenv('PLATFORM_APP_DIR') is not None):
-        STATIC_ROOT = os.path.join(os.getenv('PLATFORM_APP_DIR'), 'static')
-    if (os.getenv('PLATFORM_PROJECT_ENTROPY') is not None):
-        SECRET_KEY = os.getenv('PLATFORM_PROJECT_ENTROPY')
-    # Database service configuration, post-build only.
-    if (os.getenv('PLATFORM_ENVIRONMENT') is not None):
-        platformRelationships = decode(os.getenv('PLATFORM_RELATIONSHIPS'))
-        db_settings = platformRelationships[PLATFORMSH_DB_RELATIONSHIP][0]
-        DATABASES = {
-            'default': {
-                'ENGINE': 'django.db.backends.postgresql',
-                'NAME': db_settings['path'],
-                'USER': db_settings['username'],
-                'PASSWORD': db_settings['password'],
-                'HOST': db_settings['host'],
-                'PORT': db_settings['port'],
-            },
-            'sqlite': {
-                'ENGINE': 'django.db.backends.sqlite3',
-                'NAME': os.path.join(BASE_DIR, 'db.sqlite3'),
-            }
-        }
